nearshore/weatherapi/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.6
Markdown==2.6.9
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
urllib3==1.22
